move winforms project sources to the root folder of the project
