+++ TicTacToe.WinForms/GamePanel.cs
