--- TicTacToe.WinForms/Krest/Krest/app.config
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <configSections>
-    </configSections>
-    <connectionStrings>
-        <add name="GamePanelApplication.Properties.Settings.Database1ConnectionString"
-            connectionString="Data Source=|DataDirectory|\Database1.sdf"
-            providerName="Microsoft.SqlServerCe.Client.3.5" />
-    </connectionStrings>
-</configuration>
