+++ TicTacToe.WinForms/Properties/Resources.resx
