+++ TicTacToe.WinForms/Properties/Settings.settings
