+++ TicTacToe.WinForms/Resources/White.jpg
